btrfs-progs: handle EINVAL when reading zone size on older kernels
A combination of new progs and old kernel may lead to problems with
detecting zone size by ioctl. Fixed by #376 but still incomplete because
old kernels may return EINVAL for unsupported ioctl. This should be
ENOTTY but hasn't been like that until kernel 5.11.

As we always pass valid arguments to the ioctl we can't conflate the two
and can EINVAL the same way as ENOTTY.

Issue: #399
Signed-off-by: David Sterba <dsterba@suse.com>